Delete luckySevens.js
This was an earlier version of the other file in the directory.  It was erroneously uploaded but MUST be removed because the project is to be analyzed by a hiring manager.  It would not go so well if they analyzed the earlier, incomplete, and bug infested file instead of my final deliverable.